feat: citations, table list, mentions
