+++ packages/nocodb-sdk/src/lib/v3/index.ts
@@ -1,0 +1,30 @@
+export {
+  ApiV3DataTransformationBuilder,
+  builderGenerator,
+  colOptionBuilder,
+  columnBuilder,
+  columnOptionsV3ToV2Builder,
+  columnV3ToV2Builder,
+  sortBuilder,
+  filterBuilder,
+  filterRevBuilder,
+  viewColumnBuilder,
+} from './api-v3-data-transformation.builder';
+
+export type { ViewColumnTransformType } from './api-v3-data-transformation.builder';
+
+export {
+  getCompositePkValue,
+  recordV2ToV3,
+  recordsV2ToV3,
+  recordV3ToV2,
+  recordsV3ToV2,
+} from './record-transform';
+
+export type {
+  RecordColumnMeta,
+  ModelMeta,
+  DataRecordV3,
+  RecordV2ToV3Options,
+  RecordV3ToV2Options,
+} from './record-transform';
